remove hard code unit test input and move mongo server to utils
